breaking: Remove FeedHandler class
To migrate, change

```js
const handler = new FeedHandler((err) =>
    cb(err, handler.feed)
)
```

to

```js
const handler = new DomHandler((err) =>
    cb(err, getFeed(handler.dom))
);
```
